refactor: merge ChatComponent mixins
